i believe fixed everything

+++ Controllers/DepartmentsController.cs
@@ -16,193 +16,32 @@ namespace TallinnaRakenduslikCollegeTARpe24_StenUesson.Controllers
         public async Task<IActionResult> Index()
         {
             var schoolContext = _context.Departments.Include(d => d.Administrator);
-            return View(await (schoolContext.ToListAsync()));
-        }
-
-        public async Task<IActionResult> Details(int? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-            string query = "SELECT * FROM Departments WHERE DepartmentID = {0}";
-            var department = await _context.Departments
-                                            .FromSqlRaw(query, id)
-                                            .Include(d => d.Administrator)
-                                            .AsNoTracking()
-                                            .FirstOrDefaultAsync();
-            if (department == null)
-            {
-                return NotFound();
-            }
-            return View(department);
+            return View(await schoolContext.ToListAsync());
         }
         [HttpGet]
         public IActionResult Create()
         {
-            ViewData["InstructorID"] = new SelectList(_context.Instructors, "ID", "FullName");
-            return View();
+            ViewData["InstructorID"] = new SelectList(_context.Instructors, "id", "FullName");
+            //ViewData["CurrentStatus"] = new SelectList(_context.Departments, "id", _context.Departments.Count());
+            //var newdep = new Department();
+            return View(/*await newdep*/);
         }
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Name, Budget, StartDate, RowVersion, InstructorID, Personality, MonthlyRevenue")] Department department)
+        public async Task<IActionResult> Create([Bind("Name,Budget,StartDate,RowVersion,InstructorID,MonthlyRevenue,Personality,DepartmentDescription")] Department department)
         {
-            if (!ModelState.IsValid)
+            if (ModelState.IsValid)
             {
                 _context.Add(department);
                 await _context.SaveChangesAsync();
                 return RedirectToAction("Index");
             }
-            ViewData["InstructorID"] = new SelectList(_context.Instructors, "ID", "FullName", department.InstructorID);
-            return View(department);
 
+            ViewData["InstructorID"] = new SelectList(_context.Instructors, "id", "FullName", department.InstructorID);
+            // ViewData["CourseStatus"] = new SelectList(_context.Instructors, "id", department.CurrentStatus.ToString(), department.StudentID);
+            return View(department);
         }
         [HttpGet]
-        public async Task<ActionResult> Edit(int id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-            var departmentToEdit = await _context.Departments
-                    .Include(i => i.Administrator)
-                    .AsNoTracking()
-                    .FirstOrDefaultAsync(m => m.DepartmentID == id);
-            if (departmentToEdit == null) { return NotFound(); }
-            ViewData["InstructorID"] = new SelectList(_context.Instructors, "ID", "FullName", departmentToEdit.InstructorID);
-            return View(departmentToEdit);
-        }
-
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, byte[] rowVersion)
-        {
-            ModelState.Remove("rowVersion");
-            if (id == null) { return NotFound(); }
-            var departmentToUpdate = await _context.Departments.Include(i => i.Administrator)
-                .FirstOrDefaultAsync(m => m.DepartmentID == id);
-            if (departmentToUpdate == null)
-            {
-                Department departmentIsDeleted = new Department();
-                await TryUpdateModelAsync(departmentIsDeleted);
-                ModelState.AddModelError(string.Empty, "unable to save changes. Department has already been removed.");
-                ViewData["InstructorID"] = new SelectList(_context.Instructors, "ID", "LastName", departmentIsDeleted.InstructorID);
-                return View(departmentIsDeleted);
-            }
-            _context.Entry(departmentToUpdate).Property("RowVersion").OriginalValue = rowVersion;
-
-            var tryUpdate = await TryUpdateModelAsync<Department>(departmentToUpdate,
-                "",
-                s => s.Name,
-                s => s.StartDate,
-                s => s.Budget,
-                s => s.DepartmentDescription,
-                s => s.InstructorID,
-                s => s.Personality,
-                s => s.MonthlyRevenue);
-
-            if (tryUpdate)
-            {
-                try
-                {
-                    await _context.SaveChangesAsync();
-                    return RedirectToAction("Index");
-                }
-                catch (DbUpdateConcurrencyException ex)
-                {
-                    var exceptionEntry = ex.Entries.Single();
-                    var clientValues = (Department)exceptionEntry.Entity;
-                    var databaseEntry = exceptionEntry.GetDatabaseValues();
-
-                    if (databaseEntry == null)
-                    {
-                        ModelState.AddModelError(string.Empty, "unable to save changes, department has already been removed.");
-                    }
-                    else
-                    {
-                        var databaseValues = (Department)databaseEntry.ToObject();
-                        if (databaseValues.Name != clientValues.Name) { ModelState.AddModelError("Name", $"Current value: {databaseValues.Name}"); }
-                        if (databaseValues.StartDate != clientValues.StartDate) { ModelState.AddModelError("Name", $"Current value: {databaseValues.StartDate}"); }
-                        if (databaseValues.Budget != clientValues.Budget) { ModelState.AddModelError("Name", $"Current value: {databaseValues.Budget}"); }
-                        if (databaseValues.Personality != clientValues.Personality) { ModelState.AddModelError("Name", $"Current value: {databaseValues.Personality}"); }
-                        if (databaseValues.MonthlyRevenue != clientValues.MonthlyRevenue) { ModelState.AddModelError("Name", $"Current value: {databaseValues.MonthlyRevenue}"); }
-                        if (databaseValues.DepartmentDescription != clientValues.DepartmentDescription) { ModelState.AddModelError("Name", $"Current value: {databaseValues.DepartmentDescription}"); }
-                        if (databaseValues.InstructorID != clientValues.InstructorID) { ModelState.AddModelError("Name", $"Current value: {databaseValues.InstructorID}"); }
-                        {
-                            Instructor databaseHasThisInstructor = await _context.Instructors.FirstOrDefaultAsync(i => i.ID == databaseValues.InstructorID);
-                            ModelState.AddModelError("Name", $"Current value: {databaseValues.InstructorID}");
-                        }
-                        ModelState.AddModelError(string.Empty, "warning, changes you are about to save differ from the info in the DB" + "It appears this department was already" +
-                            "changed after you selected the version with the old info." +
-                            "click back if this new info is already correct, otherwise, click save again to oversave the department anyways.");
-                        departmentToUpdate.RowVersion = databaseValues.RowVersion;
-                        ModelState.Remove("RowVersion");
-                    }
-                }
-            }
-            ViewData["InstructorID"] = new SelectList(_context.Instructors, "ID", "Fullname", departmentToUpdate.InstructorID);
-            return View(departmentToUpdate);
-        }
-
-        //[HttpGet]
-        //public async Task<IActionResult> BaseOn(int? id)
-        //{
-        //    var department = _context.Departments.Find(id);
-        //    return View(department);
-        //}
-
-        //[HttpPost]
-        //public async Task<IActionResult> BaseOn(Department department)
-        //{
-        //    if (ModelState.IsValid)
-        //    {
-        //        var baseonDepartment = department;
-        //        _context.Departments.Add(baseonDepartment);
-        //        await _context.SaveChangesAsync();
-        //        return RedirectToAction(nameof(Index));
-        //    }
-        //    return View(department);
-        //}
-
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<ActionResult> Make([Bind("InstructorID,Name,Budget,StartDate,Personality,MonthlyRevenue")] Department department)
-        {
-            _context.Add(department);
-            await _context.SaveChangesAsync();
-            return RedirectToAction("Index");
-        }
-
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<ActionResult> MakeDelete([Bind("InstructorID,Name,Budget,StartDate,Personality,MonthlyRevenue")] Department department)
-        {
-            if (ModelState.IsValid)
-            {
-                var existingDepartment = await _context.Departments.FindAsync(department.DepartmentID);
-
-                if (existingDepartment == null)
-                {
-                    return NotFound();
-                }
-                var departmentCopy = new Department
-                {
-                    Name = department.Name,
-                    Budget = department.Budget,
-                    StartDate = department.StartDate,
-                    Personality = department.Personality,
-                    MonthlyRevenue = department.MonthlyRevenue,
-                    InstructorID = department.InstructorID
-                };
-
-                _context.Remove(existingDepartment);
-                _context.Add(departmentCopy);
-                await _context.SaveChangesAsync();
-                return RedirectToAction("Index");
-            }
-            return View(department);
-        }
-
         public async Task<IActionResult> Delete(int? id)
         {
             if (id == null)
@@ -210,22 +49,31 @@ namespace TallinnaRakenduslikCollegeTARpe24_StenUesson.Controllers
                 return NotFound();
             }
             var department = await _context.Departments
-                .FirstOrDefaultAsync(m => m.DepartmentID == id);
+                .Include(d => d.Administrator)
+                .FirstOrDefaultAsync(d => d.DepartmentID == id);
             if (department == null)
             {
                 return NotFound();
             }
             return View(department);
         }
-
-        [HttpPost, ActionName("Delete")]
+        [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
+        public async Task<IActionResult> Delete(Department department)
         {
-            var department = await _context.Departments.FindAsync(id);
-            _context.Departments.Remove(department);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            if (await _context.Departments.AnyAsync(m => m.DepartmentID == department.DepartmentID))
+            {
+                _context.Departments.Remove(department);
+                await _context.SaveChangesAsync();
+            }
+            return RedirectToAction("Index");
         }
+        [HttpGet]
+        public async Task<IActionResult> Details(int? id)
+        {
+            ViewData["Edit"] = new SelectList(_context.Instructors, "id", "FullName");
+            return View();
+        }
+
     }
 }
